+++ src/math-preview.js
@@ -18,31 +18,44 @@ function activate(context) {
 // initial preview Array
 let decorationArray = new Array;
 
+// handle MathJax error. Reload on error once.
+let onError = false;
+let resetError = false;
+
 function setPreview() {
-    
+
     clearPreview()
+
+    if (vscode.window.activeTextEditor === undefined) return;
+
+    // handle error
+    if(onError){
+        resetError && (onError = !onError)
+        resetError = !resetError
+    }
 
     const document = vscode.window.activeTextEditor.document;
     const position = vscode.window.activeTextEditor.selection.active;
 
-    // TODO: reload hscopes when error
+    // TODO: exit when document === undefined (?)
+    // TODO: forcibly reload hscopes when error
 
-    const testScope = getMathScope(document,position)
+    const testScope = hscopes.getMathScope(document, position)
 
     // exclude (not math environment) or (in math delimiter)
     if (!!testScope && !testScope.isInBeginDelimiter && !testScope.isInEndDelimiter) {
         // display math
-        if (testScope.isDisplayMath ) {
+        if (testScope.isDisplayMath) {
             const mathRange = delimiter.getMathRange(document, position, delimiter.beginDisplayMath, delimiter.endDisplayMath)
             let mathExpression = document.getText(mathRange);
-            
+
             // don't render blank formula
-            if (mathExpression === '' || mathExpression.replace(/[\n\r ]/g, '') === '') return ; 
+            if (mathExpression === '' || mathExpression.replace(/[\n\r ]/g, '') === '') return;
 
             // get rid of "blockquote" & "list"
 
-            if (testScope.scope.indexOf("quote") !== -1){
-                mathExpression = mathExpression.replace(/[\n\r]([ \s]*>)+/g,"")
+            if (testScope.scope.indexOf("quote") !== -1) {
+                mathExpression = mathExpression.replace(/[\n\r]([ \s]*>)+/g, "")
             }
 
             // enable render \\ as line break
@@ -53,13 +66,13 @@ function setPreview() {
             pushPreview(mathExpression, previewPosition)
         }
         // inline math
-        else if (!testScope.isDisplayMath ) {
+        else if (!testScope.isDisplayMath) {
             const mathRange = delimiter.getMathRange(document, position, delimiter.beginInlineMath, delimiter.endInlineMath)
             let mathExpression = document.getText(mathRange)
 
             // don't render blank formula
             if (mathExpression === '' || mathExpression.replace(/ /g, '') === '') return;
-    
+
             const beginInfo = delimiter.jumpToBeginPosition(document, position, delimiter.beginInlineMath)
             const previewPosition = beginInfo.insertPosition
             pushPreview(mathExpression, previewPosition)
@@ -78,24 +91,27 @@ function pushPreview(mathExpression, previewPosition) {
             }
         })
         .then((MathJax) => {
-            const svg = MathJax.tex2svg(mathExpression, {
-                display: false
-            });
+            const svg = MathJax.tex2svg(mathExpression, {display: false });
             const svgString = MathJax.startup.adaptor.outerHTML(svg);
-
-                if (svgString.indexOf("error") !== -1) return
-
+            // exclude error case
+            if (svgString.indexOf("error") !== -1) return
+            // create preview panel
             let mathPreview = createPreview(svgString)
             decorationArray.push(mathPreview)
-
             vscode.window.activeTextEditor.setDecorations(mathPreview, [new vscode.Range(previewPosition, previewPosition)])
-
         })
         .catch((err) => {
             console.log(err.message);
-
-            // TODO: recall setPreview() to reload preview. beware of infinite loop.
-
+            if (onError) {
+                onError = false;
+                resetError = false;
+                return
+            }
+            else{
+                onError = true;
+                setPreview();
+                return
+            }
         });
 }
 
@@ -105,7 +121,7 @@ function createPreview(mathString) {
     const defaultCss = objectToCssString({
         // TODO: is there a better way to process svg string? 
         // Another way to escape \" is : replace(/"/g, "\\\"")
-        ['content']: `url("data:image/svg+xml;utf8,${mathString.replace(/#/g,"%23").replace(/"/g,"'").replace(/<mjx-container[^<]*><svg/,"<svg").replace("</mjx-container>","")}")`,
+        ['content']: `url("data:image/svg+xml;utf8,${mathString.replace(/#/g, "%23").replace(/"/g, "'").replace(/<mjx-container[^<]*><svg/, "<svg").replace("</mjx-container>", "")}")`,
         ['position']: 'absolute',
         ['padding']: '0.5em',
         ['top']: `1.1rem`,
@@ -114,7 +130,8 @@ function createPreview(mathString) {
         ['pointer-events']: 'none',
         ['background-color']: '#e1e1e1',
         ['filter']: 'invert(100%)',
-        ['border']: '1px solid #5e5e5e',
+        ['border']: '0.5px solid #5e5e5e'
+        // TODO: set scale
     })
 
     return vscode.window.createTextEditorDecorationType({
@@ -138,64 +155,6 @@ function clearPreview() {
 }
 
 
-
-function getMathScope(document, position) {
-	const langId = document.languageId;
-    let matchDisplayMath    = new RegExp;
-    let matchBeginDelimiter = new RegExp;
-    let matchEndDelimiter   = new RegExp;
-    
-
-    // get Language
-    
-    if (langId === 'markdown') {
-        matchDisplayMath    = /math\.block|math\.display/
-        matchBeginDelimiter = /definition\.math\.begin/
-        matchEndDelimiter   = /definition\.math\.end/
-    }
-    else if (langId === 'latex') {
-        matchDisplayMath    = /math\.block\.environment/
-        matchBeginDelimiter = /definition\.string\.begin/
-        matchEndDelimiter   = /definition\.string\.end/
-    }
-    else{ return undefined }
-
-    // get scope 
-
-    const scope = hscopes.getScope(document, position).toString();
-
-    if ( !scope || scope.indexOf('math') === -1) {
-        return undefined
-    }
-
-    // get isDisplayMath
-
-    const isDisplayMath = (scope.match(matchDisplayMath) !== null ) ;
-
-    // get delimiter info
-
-    let isInBeginDelimiter = false;
-    let isInEndDelimiter = false;
-
-    if ( scope.match(matchBeginDelimiter) !== null ){
-        isInBeginDelimiter = true
-    }
-    else if (scope.match(matchEndDelimiter) != null) {
-        const localMatch = document.getText(new vscode.Range(position, document.lineAt(position).range.end))
-        const matchArray = isDisplayMath ? delimiter.endDisplayMath : delimiter.endInlineMath
-        isInEndDelimiter = !!(delimiter.searchSubStr(matchArray,localMatch).matchIndex)
-    }
-    // TODO for markdown-all-in-one scope!!!!!
-
-
-    return {
-        scope,
-        isDisplayMath,
-        isInBeginDelimiter,
-        isInEndDelimiter
-    }
-}
-
 function objectToCssString(settings) {
     let value = ''
     const cssString = Object.keys(settings)
@@ -206,10 +165,8 @@ function objectToCssString(settings) {
             }
         })
         .join(' ')
-
     return cssString
 }
-
 
 module.exports = {
     activate